admin/users modification - 1. adding "delete" panel 2. adding message & errorMessage fields 3. setUserIdToDelete javascript function creation
AdminController
adding @RequestMapping(GET) handling user deletion for admin role

========JPA reparation============
User Repository fixing:
1. changing int input variables to long

UserService & UserServiceImpl:
1. changing int input variables to long

+++ src/main/java/com/example/repository/UserRepository.java
@@ -17,7 +17,7 @@ public interface UserRepository extends JpaRepository<User, Integer> {
 	 @Transactional
 	 @Modifying
 	 @Query("DELETE from User u where u.id=?")
-	 void removeUserById(int id);
+	 void removeUserById(long id);
 	 User findById(int id);
